crypto-intel-rag/requirements.txt
streamlit>=1.32.0
google-genai>=0.5.0
sentence-transformers>=2.7.0
endee>=0.1.9
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
numpy>=1.26.0,<2.0
pandas>=2.2.0
plotly>=5.20.0
tiktoken>=0.6.0
